Move SimpleCondition to service package, as it is a private peer class to SegmentRunner.
This removes the junk (*cough* utils) package.
